requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.4
mss==9.0.1
Pillow==10.0.0
pynput==1.7.6
python-socketio==5.8.0
